Using smaller indices

+++ src/chunk.h
@@ -106,7 +106,7 @@ private:
 
     // Opaque vertices
     std::vector<BlockVertex> _opaque_vertices;
-    std::vector<uint32_t> _opaque_indices;
+    std::vector<uint16_t> _opaque_indices;
     size_t _reserved_opaque_vertex_count = 0;
     GLuint _opaque_vao;
     GLuint _opaque_vbo;
@@ -114,7 +114,7 @@ private:
 
     // Transparent vertices
     std::vector<BlockVertex> _transparent_vertices;
-    std::vector<uint32_t> _transparent_indices;
+    std::vector<uint16_t> _transparent_indices;
     size_t _reserved_transparent_vertex_count = 0;
     GLuint _transparent_vao;
     GLuint _transparent_vbo;
